+++ frontend/src/src/components/Navbar.js
@@ -4,6 +4,7 @@ import { useState } from 'react';
 
 function Navbar() {
         const handleButtonClick = () => {};
+        const [version,] = useState("1.5.4")
       
         const [showModal, setShowModal] = useState(false);
       
@@ -23,7 +24,7 @@ function Navbar() {
   return (
     <div>
          <nav className="navbar navbar-dark bg-primary justify-content-between ">
-          <a className="navbar-brand" href="/"> &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Grader</a>
+          <a className="navbar-brand" href="/"> &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Grader <a style={{fontSize: "0.7em", color: "rgb(203, 203, 203)"}}>{version}</a></a>
           <form className="form-inline">
           <Link to="/">
               <button onClick={handleButtonClick} className="btn btn-outline-light" type="button">Home