+++ frontend/src/src/pages/ST/Lab.js
@@ -227,6 +227,55 @@ function Lab() {
       })
     }
   }
+
+  const downall = async () => {
+    fetch(`${process.env.REACT_APP_HOST}/ST/assignment/downloadZip`, {
+      method: 'POST',
+      credentials: "include",
+      headers: {
+          "Content-type": "application/json; charset=UTF-8",
+          "Access-Control-Allow-Origin": "*",
+          "X-CSRF-TOKEN": Cookies.get("csrf_token")
+      },
+        body: JSON.stringify({ LID: LID})
+    })
+    .then(response => response.json())
+    .then(data => {
+      if(data.success){
+        // Decode base64-encoded file content
+        const decodedFileContent = atob(data.fileContent);
+
+        // Convert decoded content to a Uint8Array
+        const arrayBuffer = new Uint8Array(decodedFileContent.length);
+        for (let i = 0; i < decodedFileContent.length; i++) {
+            arrayBuffer[i] = decodedFileContent.charCodeAt(i);
+        }
+
+        // Create a Blob from the array buffer
+        const blob = new Blob([arrayBuffer], { type: data.fileType });
+
+        // Create a temporary URL to the blob
+        const url = window.URL.createObjectURL(blob);
+
+        // Create a link element to trigger the download
+        const a = document.createElement('a');
+        a.style.display = 'none';
+        a.href = url;
+        a.download = data.downloadFilename;
+        document.body.appendChild(a);
+        a.click();
+
+        // Clean up by revoking the object URL
+        window.URL.revokeObjectURL(url);
+      }else{
+        withReactContent(Swal).fire({
+          title: data.msg,
+          icon: "error"
+        })
+      }
+    })
+    .catch(error => console.error('Error:', error));
+  }
   
   return (
     <div>
@@ -293,7 +342,14 @@ function Lab() {
               {LabInfo.Info["Access"] ? (
               <div className='card'>
                 <div className='card-header'>
-                  <h5><Download /> Downlaod files</h5>
+                  <div className='row'>
+                    <div className='col'>
+                      <h5><Download /> Downlaod files</h5>
+                    </div>
+                    <div className='col-4'>
+                    <button type="button" className="btn btn-outline-dark" onClick={downall}>Download All</button>
+                    </div>
+                  </div>
                 </div>
                 <div className='card-body'>
                   {LabInfo.Question.map((q, i) => {